Short is replaced with Int

+++ src/main/scala/ru/ispras/modis/tm/documents/Document.scala
@@ -16,13 +16,13 @@ import scala.collection.mutable
  * @param attributes map from attribute to words, corresponding to this attribute
  * @param serialNumber serial number of document in collection
  */
-class Document(private[documents] val attributes: Map[AttributeType, Seq[(Int, Short)]], val serialNumber: Int) extends DefaultChecker {
+class Document(private[documents] val attributes: Map[AttributeType, Seq[(Int, Int)]], val serialNumber: Int) extends DefaultChecker {
     /**
      * return words, corresponding to given attribute
      * @param attributeType type of attribute
      * @return array (index of word, number of occurrence in given document
      */
-    def getAttributes(attributeType: AttributeType): Seq[(Int, Short)] = attributes.getOrElse(attributeType, Seq[(Int, Short)]())
+    def getAttributes(attributeType: AttributeType): Seq[(Int, Int)] = attributes.getOrElse(attributeType, Seq[(Int, Int)]())
 
     /**
      *
